feat: Design changes and refactoring
Design changes: Responsive index
Refactoring: Adding configuration script
