+++ public/stylesheets/index.sass
@@ -35,13 +35,16 @@ main
       list-style-type: none
 
       li
+        color: #03414C
+        font-weight: bolder
         font-size: 1em
         padding: 2em 0 2.4em 8em
         max-width: 28em
         margin: auto
-        background-repeat: no-repeat
-        background-size: 4em
-        background-position: left center
+        background:
+          repeat: no-repeat
+          size: 4em
+          position: left center
 
         h2
           font-weight: 700
