+++ server.js
@@ -5,29 +5,19 @@
 var app = require('./app')
 var debug = require('debug')('laundree:server')
 var http = require('http')
-var socketIo = require('socket.io')
-var redis = require('socket.io-redis')
+var config = require('./config')
+var socketIo = require('./src/socket-io')
 /**
  * Get port from environment and store in Express.
  */
 
-var port = normalizePort(process.env.PORT || '3000')
+var port = normalizePort(config.web.port)
 app.set('port', port)
 
 /**
  * Create HTTP server.
  */
 var server = http.createServer(app)
-
-var io = socketIo(server)
-
-if (process.env.REDIS_HOST && process.env.REDIS_PORT) {
-  var config = {
-    host: process.env.REDIS_HOST,
-    port: process.env.REDIS_PORT
-  }
-  io.adapter(redis(config))
-}
 
 /**
  * Listen on provided port, on all network interfaces.
@@ -36,12 +26,7 @@ module.exports.start = function () {
   server.listen(port)
   server.on('error', onError)
   server.on('listening', onListening)
-  // Socket.io
-  io.on('connection', (socket) => {
-    socket.on('message', (data) => {
-      io.sockets.emit('message', data)
-    })
-  })
+  socketIo(server)
 }
 
 /**
