+++ views/layout.hbs
@@ -33,17 +33,18 @@
     {{#unless no_top}}
         <nav>
             <a href="/" class="home"> </a>
-            <a href="/about">Who are we?</a>
-            <a href="#">Contact us.</a>
+            <a href="/about" class="icon about">About us</a>
+            <a href="/help" class="icon help">Help</a>
+            <a href="/contact" class="icon contact">Contact</a>
             <a href="/sign-in" class="sign-in">Sign in</a>
-            <div class="language">
+            <!--<div class="language">
                 <span>English</span>
                 <ul>
                     <li><a href="#">English</a></li>
                     <li><a href="#">Danish</a></li>
                 </ul>
 
-            </div>
+            </div>-->
         </nav>
         {{#if compact_top }}
             <header class="compact">
